The main tests run at good speed
I need some better stress tests. But so far it looks like everything is
running well.

There is a somewhat hidden bug. Sometimes the cursor is being set to an
invalid address. This is caught and the cursor is reset. But it should
not be incorrectly set in the first place.

Performance is good. Allocation times are around 100 ns + 0.1 ns per
byte. These numbers come from averaging over a loop. This can be
improved. Though the main performance issue is in the lack of proper
defragmentation and freeing of unnecessary volumes. Freed memory will be
reused, but only if sufficiently long contiguous blocks are available.